🐞 fix: 修正空闲等待与调度状态处理

+++ src/port/xf_task_port.c
@@ -77,12 +77,12 @@ void xf_task_context_swap(xf_task_manager_t manager, void *old_context, void *ne
 
 #endif // XF_TASK_CONTEXT_IS_ENABLE
 
-int32_t xf_task_msec_to_ticks(int32_t msec)
+int64_t xf_task_msec_to_ticks(int64_t msec)
 {
     return (msec * XF_TASK_TICKS_FREQUENCY / 1000);
 }
 
-int32_t xf_task_ticks_to_msec(int32_t ticks)
+int64_t xf_task_ticks_to_msec(int64_t ticks)
 {
     return (ticks * 1000 / XF_TASK_TICKS_FREQUENCY);
 }
